Enable/Disable Satchels inventory sections based on equipped items correctly
The satchel row will only appear if you have an item in the "satchel" equipment
slot, etc.
The changes are (hopefully) reflected in both the client and the server
correctly.

I implemented this by having all slots always present, and disabling the ones
the player shouldn't be able to access. I found this to be an easier approach
than implementing a container with a variable size; doing that would require
careful syncing of the container size between the client and the server to avoid
any out of bounds exceptions.