Make InventorySimpleNotifying only notify when contents really change
Hopefully fixes desync that happens when the Satchels equipment GUI is opened
when TrashSlot is present. It made redoSlots() only get called on the client,
since opening this GUI always made updateInventories() get called on the client,
while the method only got called on the server when the slot contents actually
change. Now it should get called on the client only in that case, too.